Clean up tests to move expensive test into nocover
